Refactor ProfileModel and ProfileUpdateForm, add new tests
Type: refactor, test

+++ src/apps/profiles/models.py
@@ -1,25 +1,33 @@
 import os
 
 from django.contrib.auth import get_user_model
+from django.core.exceptions import ValidationError
 from django.db import models
 from django.urls import reverse
 
 from .apps import ProfilesConfig
 
 
+def get_photo_upload_path(instance, filename):
+    ''' Profile photo will be uploaded to this path. '''
+    return os.path.join('profiles', str(instance.user.id), 'photos', filename)
+
+
 class Profile(models.Model):
 
-    # Constants
+    error_messages = {
+        'invalid_delete_photo': 'Can not delete default photo.',
+    }
+
     PHOTO_MAX_WIDTH, PHOTO_MAX_HEIGHT = 250, 250
-    DEFAULT_PHOTO_PATH =  os.path.join('profiles', 'default-avatar.jpg')
+    DEFAULT_PHOTO_PATH = os.path.join('profiles', 'default-avatar.jpg')
     MALE, FEMALE = 'M', 'F'
     GENDER_CHOICES = ((MALE, 'Male'), (FEMALE, 'Female'))
 
-    # Helper function.
-    def get_photo_upload_path(profile, filename):
-        return os.path.join('profiles', str(profile.user.id), 'photos', filename)
+    @staticmethod
+    def get_max_photo_size_display():
+        return f'{Profile.PHOTO_MAX_WIDTH} x {Profile.PHOTO_MAX_HEIGHT}'
 
-    # Fields
     user = models.OneToOneField(get_user_model(), on_delete=models.CASCADE)
     first_name = models.CharField(max_length=128, default='', blank=True)
     last_name = models.CharField(max_length=128, default='', blank=True)
@@ -32,17 +40,16 @@ class Profile(models.Model):
     age = models.PositiveSmallIntegerField(default=0, blank=True)
     login_count = models.IntegerField(default=0)
 
-    # Python methods
     def __str__(self):
         return self.user.username
 
     def __repr__(self):
         return (
-            f'Profile(first_name={self.first_name}, last_name={self.last_name}, '
-            f'photo={self.photo}, gender={self.gender}, age={self.age})'
+            f"{self.__class__.__name__}(first_name='{self.first_name}', "
+            f"last_name='{self.last_name}', photo='{self.photo}', "
+            f"gender='{self.gender}', age={self.age})"
         )
 
-    # Custom methods
     def get_absolute_url(self):
         return reverse(ProfilesConfig.PROFILE_DETAIL_URL, args=[self.user.username])
 
@@ -59,14 +66,14 @@ class Profile(models.Model):
         return self.gender == self.FEMALE
 
     def has_default_photo(self):
-        return self.photo == Profile.DEFAULT_PHOTO_PATH
+        return self.photo == self.DEFAULT_PHOTO_PATH
 
     def has_uploaded_photo(self):
-        return self.photo != Profile.DEFAULT_PHOTO_PATH
+        return self.photo != self.DEFAULT_PHOTO_PATH
 
-    # TODO Something bus me in here.
     def delete_photo(self):
-        ''' Delete user's profile photo. '''
+        if self.has_default_photo():
+            raise ValidationError(self.error_messages['invalid_delete_photo'])
         try:
             os.remove(self.photo.path)
         except FileNotFoundError as e:
